* develop: (81 commits)
  added callback support to interpreter's invoke method
  arm benchmark model and argmax pass bug fixed
  saving intermediate passes enabled in xfomer
  logging module renamed to avoid circular dependencies
  save_intermediate method added to pass manager, but untested
  improved log_method_output
  log_method_output moved to logging module
  new logging module and debug mode added
  updated macos libtflite2xcore
  fixing logger bug
  better verbosity handling added
  more minification passes added, option available in xformer
  moved marking of test XFAILs into json test case spec file
  XPASS integration tests now PASS, XFAILs still XFAIL
  no longer dynamically allocates par regions
  LegalizeOutputTensorNamePass renamed and enabled in conversion
  mutate tests added to LegalizeOutputTensorNamePass
  LegalizeOutputTensorNamePass matching tests completed
  LegalizeOutputTensorNamePass matching implemented and partially tested
  small simplificatin in lut_passes tests
  ...

+++ tests/test_op_models.py
@@ -9,10 +9,15 @@ import subprocess
 import tempfile
 import pytest
 
+import numpy as np
+
 import helpers
 import directories
-from tflite2xcore import read_flatbuffer
 from tflite2xcore import operator_codes
+from tflite2xcore.serialization import read_flatbuffer
+from tflite2xcore.xcore_interpreter import XCOREInterpreter
+from tflite2xcore.xcore_model import TensorType
+
 
 def load_tests(test_name, test_dir, max_count):
     supported_operators = set([
@@ -49,7 +54,7 @@ def load_tests(test_name, test_dir, max_count):
 
                 for input_file, output_file in zip(sorted(input_files), sorted(output_files)):
                     test_cases.append({
-                        'id': '/'.join(directory.parts[-2:]),
+                        'id': '/'.join(directory.parts[-2:])+'_',
                         'flatbuffer': flatbuffer_xcore,
                         'input': {
                             'filename': input_file,
@@ -84,74 +89,129 @@ def run_test_case(test_model_app, test_case, abs_tol=1):
     expected_output_file = test_case['expected_output']['filename']
     expected_quantization = test_case['expected_output']['quantization']
     predicted_output_file = os.path.join(tempfile.mkdtemp(), 'predicted_output.bin')
-    if test_model_app.endswith('.xe'):
-        cmd = f'xsim --args {test_model_app} {flatbuffer} {input_file} {predicted_output_file}'
-    else:
-        cmd = f'{test_model_app} {flatbuffer} {input_file} {predicted_output_file}'
+
     print('**********')
     print('* Inputs *')
     print('**********')
     # print('Testcase:', json.dumps(test_case, indent=4))
-    print('Command:', cmd)
     print('***********')
     print('* Results *')
     print('***********')
-    try:
-        output = subprocess.check_output(cmd, stderr=subprocess.STDOUT, shell=True)
+    if test_model_app:
+        if test_model_app.endswith('.xe'):
+            cmd = f'xsim --args {test_model_app} {flatbuffer} {input_file} {predicted_output_file}'
+        else:
+            cmd = f'{test_model_app} {flatbuffer} {input_file} {predicted_output_file}'
+        print('Command:', cmd)
+        try:
+            output = subprocess.check_output(cmd, stderr=subprocess.STDOUT, shell=True)
+            result = helpers.compare_tensor_files(expected_output_file, expected_quantization,
+                predicted_output_file, predicted_quantization, abs_tol)
+
+            os.remove(predicted_output_file)
+
+            return result
+        except subprocess.CalledProcessError as ex:
+            print(ex.output.decode('utf-8'))
+            print(ex)
+            return False
+    else:
+        # use interpreter
+        model = read_flatbuffer(flatbuffer)
+
+        input_tensor = model.subgraphs[0].inputs[0]
+        input_index= model.subgraphs[0].tensors.index(model.subgraphs[0].inputs[0])
+        output_tensor = model.subgraphs[0].outputs[0]
+        output_index= model.subgraphs[0].tensors.index(model.subgraphs[0].outputs[0])
+
+        with open(flatbuffer, 'rb') as fd:
+            model_content = fd.read()
+
+        interpreter = XCOREInterpreter(model_content=model_content)
+        interpreter.allocate_tensors()
+
+        input_ = np.fromfile(input_file, dtype=TensorType.to_numpy_dtype(input_tensor.type))
+        input_ = input_.reshape(input_tensor.shape)
+
+        interpreter.set_tensor(input_index, input_)
+        interpreter.invoke()
+        predicted_output = interpreter.get_tensor(output_index)
+
+        with open(predicted_output_file, 'wb') as fd:
+            fd.write(predicted_output.tobytes())
+
         result = helpers.compare_tensor_files(expected_output_file, expected_quantization,
             predicted_output_file, predicted_quantization, abs_tol)
 
-        os.remove(predicted_output_file)
-
         return result
-    except subprocess.CalledProcessError as ex:
-        print(ex)
-        return False
+
+def is_xfail(test_case):
+    index = test_case['id'].find('xfail=true')
+    return index > 0
 
 def test_XC_lookup_8(test_model_app, XC_lookup_8_test_case, abs_tol):
+    if is_xfail(XC_lookup_8_test_case):
+        pytest.xfail()
     assert(run_test_case(test_model_app, XC_lookup_8_test_case, abs_tol))
 
 
 def test_XC_argmax_16(test_model_app, XC_argmax_16_test_case, abs_tol):
+    if is_xfail(XC_argmax_16_test_case):
+        pytest.xfail()
     assert(run_test_case(test_model_app, XC_argmax_16_test_case, abs_tol))
 
 
 def test_XC_conv2d_1x1(test_model_app, XC_conv2d_1x1_test_case, abs_tol):
+    if is_xfail(XC_conv2d_1x1_test_case):
+        pytest.xfail()
     assert(run_test_case(test_model_app, XC_conv2d_1x1_test_case, abs_tol))
 
 
 def test_XC_conv2d_depthwise(test_model_app, XC_conv2d_depthwise_test_case, abs_tol):
+    if is_xfail(XC_conv2d_depthwise_test_case):
+        pytest.xfail()
     assert(run_test_case(test_model_app, XC_conv2d_depthwise_test_case, abs_tol))
 
 
-@pytest.mark.xfail
 def test_XC_conv2d_shallowin_deepout_relu(test_model_app, XC_conv2d_shallowin_deepout_relu_test_case, abs_tol):
+    if is_xfail(XC_conv2d_shallowin_deepout_relu_test_case):
+        pytest.xfail()
     assert(run_test_case(test_model_app, XC_conv2d_shallowin_deepout_relu_test_case, abs_tol))
 
 
-@pytest.mark.xfail
 def test_XC_conv2d_deepin_deepout_relu(test_model_app, XC_conv2d_deepin_deepout_relu_test_case, abs_tol):
+    if is_xfail(XC_conv2d_deepin_deepout_relu_test_case):
+        pytest.xfail()
     assert(run_test_case(test_model_app, XC_conv2d_deepin_deepout_relu_test_case, abs_tol))
 
 
 def test_XC_fc_deepin_anyout(test_model_app, XC_fc_deepin_anyout_test_case, abs_tol):
+    if is_xfail(XC_fc_deepin_anyout_test_case):
+        pytest.xfail()
     assert(run_test_case(test_model_app, XC_fc_deepin_anyout_test_case, abs_tol))
 
 
 def test_XC_maxpool2d(test_model_app, XC_maxpool2d_test_case, abs_tol):
+    if is_xfail(XC_maxpool2d_test_case):
+        pytest.xfail()
     assert(run_test_case(test_model_app, XC_maxpool2d_test_case, abs_tol))
 
 
-@pytest.mark.xfail
 def test_XC_avgpool2d(test_model_app, XC_avgpool2d_test_case, abs_tol):
+    if is_xfail(XC_avgpool2d_test_case):
+        pytest.xfail()
     assert(run_test_case(test_model_app, XC_avgpool2d_test_case, abs_tol))
 
 
 def test_XC_avgpool2d_global(test_model_app, XC_avgpool2d_global_test_case, abs_tol):
+    if is_xfail(XC_avgpool2d_global_test_case):
+        pytest.xfail()
     assert(run_test_case(test_model_app, XC_avgpool2d_global_test_case, abs_tol))
 
 
 def test_XC_requantize_16_to_8(test_model_app, XC_requantize_16_to_8_test_case, abs_tol):
+    if is_xfail(XC_requantize_16_to_8_test_case):
+        pytest.xfail()
     assert(run_test_case(test_model_app, XC_requantize_16_to_8_test_case, abs_tol))
